+++ package.json
@@ -3,16 +3,42 @@
     "version": "1.0.0",
     "description": "",
     "scripts": {
-      "postinstall": "cd ./server && npm install --dev",
-      "start": "cd ./server && npm start"
+        "postinstall": "cd ./server && npm install --dev",
+        "start": "cd ./server && npm start",
+        "test": "jest"
     },
-    "dependencies": {},
-    "devDependencies": {},
+    "dependencies": {
+        "express": "^4.17.2",
+        "jest": "^27.4.7",
+        "swagger-jsdoc": "^6.1.0",
+        "swagger-ui-express": "^4.3.0",
+        "ts-jest": "^27.1.3",
+        "yarn": "^1.22.17"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.16.10",
+        "@babel/preset-env": "^7.16.11",
+        "@types/jest": "^27.4.0",
+        "babel-jest": "^27.4.6",
+        "typescript": "^4.5.5"
+    },
     "author": "",
     "license": "MIT",
     "repository": "",
     "engines": {
         "node": "14.x"
+    },
+    "jest": {
+        "roots": [
+            "./server/Logic",
+            "./server/tests"
+        ],
+        "testMatch": [
+            "**/__tests__/**/*.+(ts|tsx|js)",
+            "**/?(*.)+(spec|test).+(ts|tsx|js)"
+        ],
+        "transform": {
+            "^.+\\.(ts|tsx)$": "ts-jest"
+        }
     }
-  }
-  
+}
